+++ CHANGELOG.rst
@@ -31,6 +31,7 @@ Fixed
 - **Projectroles**
     - Deprecated ``SODAR_API_*`` settings required in tests (#1495)
     - Add workaround to ``ProjectInviteCreateView`` returning 404 with category and query string (#1510)
+    - Broken tour help attachments in ``ProjectRoleView`` (#1512)
 
 
 v1.0.2 (2024-09-09)