+++ CHANGELOG.rst
@@ -23,6 +23,8 @@ Added
 Changed
 -------
 
+- **General**
+    - Update dependency pinning (#1509)
 - **Projectroles**
     - Update default OIDC login button template (#1503)
     - Update ownership transfer timeline event data (#1514)