+++ CHANGELOG.rst
@@ -5,8 +5,8 @@ Changelog for the **SODAR Core** Django app package. Loosely follows the
 `Keep a Changelog <http://keepachangelog.com/en/1.0.0/>`_ guidelines.
 
 
-Unreleased
-==========
+v1.0.3 (2024-12-12)
+===================
 
 Added
 -----
@@ -14,6 +14,7 @@ Added
 - **Projectroles**
     - Info link for finder role in ``ProjectRoleView`` (#1511)
     - Table and strikethrough support in ``render_markdown()`` (#1272)
+    - ``sodar-markdown-content`` CSS class (#1272)
 - **Timeline**
     - User count in siteinfo stats (#1504)
     - Plugin tests (#1506)